requirements.txt
langchain
langchain-openai
langchain-google
pydantic
python-dotenv
langgraph
wikipedia
langchain-community
#langchain_google_genai
python-dotenv

langchain_redis